factor proscenium.display into sub-packages
